requirements.txt
numpy
pandas
tensorflow

click
requests
biopython

git+git://github.com/smsaladi/sgidspace

